requirements.txt
googletrans==4.0.0-rc1
gTTS==2.5.1
pymongo==4.6.3
python-dotenv==1.0.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-telegram-bot==22.0
requests==2.31.0
transformers==4.40.0
torch==2.2.2
httpx==0.28.1